Move middleware/auth to api/auth
Also remove defunct Authorization code that's not currently used
