+++ administration/server/src/api/flags/handlers.ts
@@ -1,7 +1,7 @@
 import { HttpApi, HttpApiBuilder } from '@effect/platform';
 import { Effect, Schema } from 'effect';
+import { CurrentUser } from '~/api/auth';
 import { DatabaseLive, DatabaseService } from '~/layers/db';
-import { CurrentUser } from '~/middleware/auth';
 import { flagsGroup } from './endpoints';
 import { MemberFlagSchema } from './schemas';
