+++ administration/server/src/api/index.ts
@@ -8,7 +8,7 @@ import { Layer } from 'effect';
 import type { Effect, Sink, Stream, Channel } from 'effect';
 import type { NodeInspectSymbol } from 'effect/Inspectable';
 
-import { AuthMiddleware, AuthMiddlewareLive } from '~/middleware/auth';
+import { AuthMiddleware, AuthMiddlewareLive } from '~/api/auth';
 import { AuthLive } from '~/layers/auth';
 import { DatabaseLive } from '~/layers/db';
