+++ administration/server/src/index.ts
@@ -10,8 +10,8 @@ import dotenv from 'dotenv';
 import { Effect, Layer } from 'effect';
 import express from 'express';
 import { ApiLive } from '~/api';
+import { auditMiddleware } from '~/api/auditLogging';
 import { Auth, AuthLive } from '~/layers/auth';
-import { auditMiddleware } from '~/middleware/auditLogging';
 
 dotenv.config();
